consumerHealtCheckService/app/requirements.txt
Flask==2.2.5
kafka-python==2.0.1